libs/module: Support LoRa868v2, 470, 915.
