+++ docs/en/refs/module.lorawan868.ref
@@ -1,0 +1,62 @@
+
+.. |LoRaWAN868Module| image:: https://static-cdn.m5stack.com/resource/docs/products/module/comx_lorawan/comx_lorawan_01.webp
+    :target: https://docs.m5stack.com/en/module/comx_lorawan
+    :height: 200px
+    :width: 200px
+
+.. |init.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/init.png
+.. |set_mode.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_mode.png
+.. |set_parameters.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_parameters.png
+.. |wake_up.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/wake_up.png
+.. |sleep.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/sleep.png
+.. |reset.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/reset.png
+.. |restore_factory_settings.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/restore_factory_settings.png
+.. |set_copyright.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_copyright.png
+.. |set_auto_low_power.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_auto_low_power.png
+.. |query_chip_id.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/query_chip_id.png
+.. |enable_rx.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/enable_rx.png
+.. |set_deveui.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_deveui.png
+.. |set_appeui.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_appeui.png
+.. |set_appkey.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_appkey.png
+.. |set_nwkskey.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_nwkskey.png
+.. |set_appskey.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_appskey.png
+.. |set_devaddr.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_devaddr.png
+.. |set_otaa_mode.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_otaa_mode.png
+.. |set_adr.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_adr.png
+.. |set_channel_mask.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_channel_mask.png
+.. |join_network.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/join_network.png
+.. |set_duty_cycle.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_duty_cycle.png
+.. |set_class_mode.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_class_mode.png
+.. |set_ack.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_ack.png
+.. |set_app_port.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_app_port.png
+.. |set_retransmission_count.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/set_retransmission_count.png
+.. |send_hex.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/send_hex.png
+.. |send_string.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/send_string.png
+.. |query_lorawan_mode.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/query_lorawan_mode.png
+.. |save_parameters_to_flash.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/save_parameters_to_flash.png
+.. |at_cmd.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/at_cmd.png
+.. |at_query.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/at_query.png
+.. |at_receive.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/at_receive.png
+.. |flush.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/flush.png
+.. |any.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/any.png
+.. |receive_data.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/receive_data.png
+
+.. |example.png| image:: https://static-cdn.m5stack.com/mpy_docs/module/lorawan868module/example.png
+
+.. |lorawan868_example_tx.m5f2| raw:: html
+
+    <a
+        href="https://uiflow2.m5stack.com/?example=https://raw.githubusercontent.com/m5stack/uiflow-micropython/develop/examples/module/lorawan868_example_tx.m5f2"
+        target="_blank"
+    >
+        lorawan868_example_tx.m5f2
+    </a>
+.. |lorawan868_example_rx.m5f2| raw:: html
+
+    <a
+        href="https://uiflow2.m5stack.com/?example=https://raw.githubusercontent.com/m5stack/uiflow-micropython/develop/examples/module/lorawan868_example_rx.m5f2"
+        target="_blank"
+    >
+        lorawan868_example_rx.m5f2
+    </a>
+
