+++ examples/module/lorawan868/lorawan868_example_tx.m5f2
@@ -1,0 +1,1 @@
+{"version":"V2.0","versionNumber":"V2.1.6","type":"basic","components":[{"name":"screen","type":"screen","layer":0,"screenId":"builtin","screenName":"","id":"__basic_screen","createTime":1730346281094,"x":0,"y":0,"width":320,"height":240,"backgroundColor":"#222222","size":0,"isSelected":true}],"resources":[{"hardware":["hardware_button","hardware_pin_button","imu","rgb","speaker"]},{"module":["module_com_lorawan868"]}],"units":[],"hats":[],"bases":[],"i2cs":[],"blockly":"<block type=\"basic_on_setup\" id=\"setup_block\" deletable=\"false\" x=\"-290\" y=\"110\"><mutation isBegin=\"true\"></mutation><field name=\"UPDATEOP\">true</field><statement name=\"FUNC\"><block type=\"system_m5_begin\" id=\"system_m5_begin\"><next><block type=\"module_com_lorawan868_init\" id=\"7H;5f9+u}X_8K9Ofo$y|\"><field name=\"NAME\">lorawan868_0</field><value name=\"UART\"><shadow type=\"module_com_lorawan868_id_option\" id=\".+j,[mHxj3tTJd5}*ubU\"><field name=\"VALUE\">1</field></shadow></value><value name=\"TX\"><shadow type=\"math_number\" id=\"Wp[^#)q:=Gnw-RJ83CKi\"><mutation max=\"Infinity\" min=\"-Infinity\" precision=\"0\"></mutation><field name=\"NUM\">16</field></shadow></value><value name=\"RX\"><shadow type=\"math_number\" id=\"(+G?`M+sFL(F/]:E]8jp\"><mutation max=\"Infinity\" min=\"-Infinity\" precision=\"0\"></mutation><field name=\"NUM\">17</field></shadow></value><next><block type=\"module_com_lorawan868_wake_up\" id=\"Z`s]pycjY%6@DVKNFAq;\"><field name=\"NAME\">lorawan868_0</field><next><block type=\"module_com_lorawan868_set_parameters\" id=\"}HHs%)#XAjSiPrAEy?BY\"><field name=\"NAME\">lorawan868_0</field><value name=\"FREQ\"><shadow type=\"math_number\" id=\"tBslVQoQ($3c/+Rm8BUf\"><mutation max=\"Infinity\" min=\"-Infinity\" precision=\"0\"></mutation><field name=\"NUM\">0</field></shadow></value><value name=\"POWER\"><shadow type=\"math_number\" id=\"2YV]MO{NSY^:A3{{}UNJ\"><mutation max=\"Infinity\" min=\"-Infinity\" precision=\"0\"></mutation><field name=\"NUM\">0</field></shadow></value><value name=\"SF\"><shadow type=\"math_slider\" id=\"@OUO%$GCt+B=pnJh+`D4\"><mutation max=\"12\" min=\"5\" step=\"1\" precision=\"1\"></mutation><field name=\"NUM\">5</field></shadow></value><value name=\"BW\"><shadow type=\"module_com_lorawan868_bw_option\" id=\"eOdLxKHS4wO@D!_%=Vx=\"><field name=\"VALUE\">0</field></shadow></value><value name=\"CR\"><shadow type=\"module_com_lorawan868_cr_option\" id=\"`L$x-LX,Yi[G`so%J78D\"><field name=\"VALUE\">1</field></shadow></value><value name=\"PREAMBLE\"><shadow type=\"math_slider\" id=\"|g_+4^tw;%p=m,jZTl!P\"><mutation max=\"65535\" min=\"8\" step=\"1\" precision=\"1\"></mutation><field name=\"NUM\">8</field></shadow></value><value name=\"CRC\"><shadow type=\"module_com_lorawan868_enable_option\" id=\"4MD@hA}~W=.swj]n%{WY\"><field name=\"VALUE\">0</field></shadow></value><value name=\"IQ_INV\"><shadow type=\"module_com_lorawan868_enable_option\" id=\"DQAt;+f8+i-d)U-=%4i+\"><field name=\"VALUE\">0</field></shadow></value><value name=\"SAVE\"><shadow type=\"module_com_lorawan868_enable_option\" id=\"Y:o3_gpqXp_!c%J1V;4%\"><field name=\"VALUE\">0</field></shadow></value><next><block type=\"module_com_lorawan868_set_auto_low_power\" id=\"e*7b!-Cqv(9VH[IAUvc6\"><field name=\"NAME\">lorawan868_0</field><value name=\"VALUE\"><shadow type=\"module_com_lorawan868_bool_option\" id=\"l}E;z`PouT)BJ4)ZX[p8\"><field name=\"VALUE\">False</field></shadow></value><next><block type=\"text_print\" id=\"Mhq8GJy,aO%JplMX@%[5\"><value name=\"TEXT\"><shadow type=\"text\" id=\":/^+B|T%xi^~T[o/h)S[\"><field name=\"TEXT\">hello M5</field></shadow><block type=\"module_com_lorawan868_query_chip_id\" id=\"8C-/7qGIWySe@%X:0ldN\"><field name=\"NAME\">lorawan868_0</field></block></value><next><block type=\"text_print\" id=\",J*6tbN,{-W%xRkvA+5s\"><value name=\"TEXT\"><shadow type=\"text\" id=\"/E1Y(+A81Hws}d,T_Yql\"><field name=\"TEXT\">hello M5</field></shadow><block type=\"module_com_lorawan868_query_lorawan_mode\" id=\"]3}(4e;?8fR@+b=-[O#]\"><field name=\"NAME\">lorawan868_0</field></block></value><next><block type=\"text_print\" id=\"A%a857U_m=BQB3tvwbQ`\"><value name=\"TEXT\"><shadow type=\"text\" id=\"-^MliVfrPzUC`KzDJDYE\"><field name=\"TEXT\">hello M5</field></shadow><block type=\"module_com_lorawan868_any\" id=\"Cv.V2#?*wDBu`/RQr_@R\"><field name=\"NAME\">lorawan868_0</field></block></value><next><block type=\"module_com_lorawan868_set_mode\" id=\"98j$Go-JsrTE-c9i1M08\"><field name=\"NAME\">lorawan868_0</field><value name=\"VALUE\"><shadow type=\"module_com_lorawan868_mode_option\" id=\"f9RdKRhB{*]iAd+mg-D}\"><field name=\"VALUE\">MODE_LORA</field></shadow></value></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></next></block></statement></block><block type=\"basic_on_loop\" id=\"loop_block\" deletable=\"false\" x=\"170\" y=\"210\"><mutation isUpdate=\"true\"></mutation><field name=\"UPDATEOP\">true</field><statement name=\"FUNC\"><block type=\"system_m5_update\" id=\"system_m5_update\"><next><block type=\"module_com_lorawan868_send_hex\" id=\"N)4xRKD3^%8Oo7uHaG!v\"><field name=\"NAME\">lorawan868_0</field><value name=\"VALUE\"><shadow type=\"text\" id=\"P%Tz]j8^C4z9k0~B)s!?\"><field name=\"TEXT\">Hello Lora!</field></shadow></value><next><block type=\"time_sleep_second\" id=\"QG#gAcY1}L#hM^P~I3I2\"><value name=\"SECOND\"><shadow type=\"math_number\" id=\"gUG=1BB+*t#aCan`Po.o\"><mutation max=\"Infinity\" min=\"0\" precision=\"0\"></mutation><field name=\"NUM\">1</field></shadow></value></block></next></block></next></block></statement></block>","screen":[{"simulationName":"Built-in","type":"builtin","width":320,"height":240,"scale":0.78,"screenName":"","blockId":"","screenColorType":0,"id":"builtin","createTime":1730346281094}],"logicWhenNum":0,"customList":[]}
