+++ m5stack/libs/module/__init__.py
@@ -14,12 +14,13 @@ _attrs = {
     "IotBaseCatmModule": "iot_base_catm",
     "LoraModule": "lora",
     "LoRaWANModule": "lorawan",
+    "LoRaWAN868Module": "lorawan868",
     "Module4In8Out": "module_4in8out",
     "PLUSModule": "plus",
     "PPSModule": "pps",
     "RCAModule": "rca",
-    "Relay4Module": "relay_4",
     "Relay2Module": "relay_2",
+    "Relay4Module": "relay_4",
     "RS232Module": "rs232",
     "StepMotorDriverModule": "step_motor_driver",
 }
