+++ m5stack/libs/module/manifest.py
@@ -15,14 +15,15 @@ package(
         "iot_base_catm.py",
         "lora.py",
         "lorawan.py",
+        "lorawan868.py",
         "mbus.py",
         "module_4in8out.py",
         "module_helper.py",
         "plus.py",
         "pps.py",
         "rca.py",
-        "relay_4.py",
         "relay_2.py",
+        "relay_4.py",
         "rs232.py",
         "step_motor_driver.py",
     ),
